Add code and presentation assets to the add_asset_box
Check if the CodeAsset and/or FassetsPresentation engine is present
and add a code or presentation asset option to the add_asset_box.
Make a few changes to the css code to ensure the items in the list
are not shown side by side. This could happen if FassetPresentations
is not present.
Make the YouTube preview 400px wide as all other previews.