fix usage of rails form_for form builder
To allow the form builder code to add the multipart option to the form,
the file input field must be created using the original form object.
This propagates the form object into the local_file_form partial and
makes use of it to create the field.

Makes file upload finally work again.